Split Agent controller tests by mode
